push-to-ipfs/scripts/requirements.txt
requests==2.32.4
py-cid==0.3.0